tweaked Makefile.am because h5part/test/Makefile.am doesn't build portably
h5part/test/Makefile.am assumes that the build is happening in h5part/ and
breaks the build when using a different build directory.  I removed the
test directory from the build list to avoid this, but we must now bootstrap
autoconf/automake before building to include this new build list.